Move unmapping tests to separate file
